change actions main component into sub component
